Move MockNetwork to `tests` in `network-controller`
`MockNetwork` is a class that can be used in tests to easily mock
requests that go through NetworkController network clients.

Because the file where this class lives (`mock-network.ts`) relies on
`network-controller` types, any package that wants to use this class
needs to declare a development dependency on `network-controller`. This
is not totally obvious due to the fact that the test helper file is kept
at the root level and not inside of `network-controller` itself, and it
confuses TypeScript and its project reference resolution logic (in fact
type errors exist in `mock-network.ts` today.

This commit moves the file inside of `network-controller`, resolving the
type errors.
